Refactor collate logic
